requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.5
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
confection==0.1.5
cryptography==45.0.6
cymem==2.0.11
dashscope==1.24.2
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
fastapi==0.116.1
filetype==1.2.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-deepseek==0.1.4
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langserve==0.3.1
langsmith==0.4.16
language_data==1.3.0
lxml==6.0.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.4
murmurhash==1.0.13
mypy_extensions==1.1.0
nltk==3.9.1
numpy==2.3.2
olefile==0.47
openai==1.101.0
orjson==3.11.2
packaging==25.0
preshed==3.0.10
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.0.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.14.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy_pkuseg==1.0.1
SQLAlchemy==2.0.43
srsly==2.5.1
sse-starlette==1.8.2
starlette==0.47.3
tenacity==9.1.2
thinc==8.3.6
tiktoken==0.11.0
tqdm==4.67.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
unstructured==0.18.14
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.3
yarl==1.20.1
zh_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.8.0/zh_core_web_sm-3.8.0-py3-none-any.whl#sha256=7de3bd267176b9b2a8defb6997c1cd296da16c57b5e712f72ea44a51755421c8
zstandard==0.24.0
